force upload

+++ app/src/main/java/com/hrd/localvoice/utils/AudioUtil.kt
@@ -10,24 +10,26 @@ import com.hrd.localvoice.AppRoomDatabase
 import com.hrd.localvoice.models.Audio
 import java.io.File
 
-//import java.io.File
-
 class AudioUtil {
     companion object {
         fun convert(audio: Audio, application: Application) {
             val inputFile = audio.localFileURl
-            val outputFile = inputFile.split(".wav")[0] + ".mp3"
+            var outputFile = inputFile.split(".wav")[0] + ".mp3"
+            outputFile = outputFile.replace("u${audio.remoteImageID}", "u${audio.userId}")
             when (val rc = FFmpeg.execute("-i $inputFile $outputFile")) {
                 RETURN_CODE_SUCCESS -> {
-                    audio.localFileURl = outputFile
-                    AppRoomDatabase.databaseWriteExecutor.execute {
-                        val id =
-                            AppRoomDatabase.getDatabase(application)?.AudioDao()?.updateAudio(audio)
-                        if (id != null) {
-                            // Delete wave file
-                            val waveFile = File(inputFile)
-                            if (waveFile.isFile) {
-                                waveFile.delete()
+                    if (File(outputFile).length() / 1024 > 100) {
+                        audio.localFileURl = outputFile
+                        AppRoomDatabase.databaseWriteExecutor.execute {
+                            val id =
+                                AppRoomDatabase.getDatabase(application)?.AudioDao()
+                                    ?.updateAudio(audio)
+                            if (id != null) {
+                                // Delete wave file
+                                val waveFile = File(inputFile)
+                                if (waveFile.isFile) {
+                                    waveFile.delete()
+                                }
                             }
                         }
                     }
@@ -36,7 +38,7 @@ class AudioUtil {
                     Log.i(Config.TAG, "Command execution cancelled by user.");
                 }
                 else -> {
-                    Log.i(
+                    Log.e(
                         Config.TAG, String.format(
                             "Command execution failed with rc=%d and the output below.", rc
                         )
